requirements.txt
absl-py>=0.5.0
mujoco-py<1.50.2,>=1.50.1
numpy>=1.13.3
tensorflow==1.15.0
tensorflow-probability==0.8.0rc0
gin-config
gym==0.12.1
tf-agents==0.3.0
